requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.11
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
drf-yasg==1.21.14
frozenlist==1.8.0
git-filter-repo==2.47.0
graphviz==0.21
gunicorn==25.0.1
idna==3.11
inflection==0.5.1
mailchimp-marketing==3.0.80
multidict==6.7.1
packaging==26.0
pillow==12.2.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydot==4.0.1
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
twilio==9.10.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.22.0
